+++ tests/attributes/test_counter.py
@@ -11,9 +11,9 @@ class MyModel(sheraf.AutoModel):
     counter = sheraf.CounterAttribute(default=0)
     useless = sheraf.InlineModelAttribute(
         sheraf.InlineModel(
-            foo=sheraf.SimpleAttribute(default="bar", lazy_creation=False),
+            foo=sheraf.SimpleAttribute(default="bar", lazy=False),
         ),
-        lazy_creation=False,
+        lazy=False,
     )