+++ tests/batches/fixture1/__init__.py
@@ -8,7 +8,7 @@ class InlineModel(sheraf.InlineModel):
 class Model1(sheraf.Model):
     table = "model1_table"
 
-    simple = sheraf.SimpleAttribute(lazy_creation=False)
-    my_inline_model = sheraf.InlineModelAttribute(InlineModel, lazy_creation=False)
-    my_blob = sheraf.BlobAttribute(lazy_creation=False)
-    counter = sheraf.CounterAttribute(lazy_creation=False)
+    simple = sheraf.SimpleAttribute(lazy=False)
+    my_inline_model = sheraf.InlineModelAttribute(InlineModel, lazy=False)
+    my_blob = sheraf.BlobAttribute(lazy=False)
+    counter = sheraf.CounterAttribute(lazy=False)